convert all float members to double in the ProtoBuf messages and corresponding FileModel classes to increase precision

+++ OdbDesignLib/FileModel/Design/FeaturesFile.h
@@ -38,22 +38,22 @@ namespace Odb::Lib::FileModel::Design
 			Type type;			
 				
 			// Line
-			float xs, ys;
-			float xe, ye;
+			double xs, ys;
+			double xe, ye;
 
 			// Pad / Text
-			float x, y;
+			double x, y;
 			int apt_def_symbol_num;
-			float apt_def_resize_factor;
+			double apt_def_resize_factor;
 
 			// Arc
-			float xc, yc;
+			double xc, yc;
 			bool cw;
 
 			// Text
 			std::string font;
-			float xsize, ysize;
-			float width_factor;
+			double xsize, ysize;
+			double width_factor;
 			std::string text;
 			int version;
 
@@ -66,7 +66,7 @@ namespace Odb::Lib::FileModel::Design
 			unsigned int id;
 
 			int orient_def;
-			float orient_def_rotation;			
+			double orient_def_rotation;			
 
 			ContourPolygon::Vector m_contourPolygons;
